+++ MHS CLUBHOUSE/MHSystems/app/src/main/java/com/mh/systems/clubhouse/ui/activites/CompletedDetailActivity.java
@@ -1,6 +1,7 @@
 package com.mh.systems.clubhouse.ui.activites;
 
 import android.content.DialogInterface;
+import android.graphics.Bitmap;
 import android.graphics.Typeface;
 import android.os.Bundle;
 import android.support.design.widget.AppBarLayout;
@@ -10,18 +11,23 @@ import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ExpandableListAdapter;
+import android.widget.ExpandableListView;
 import android.widget.LinearLayout;
-import android.widget.ListView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.google.gson.JsonObject;
 import com.mh.systems.clubhouse.R;
 import com.mh.systems.clubhouse.ui.adapter.BaseAdapter.CompetitionDetailAdapter;
+import com.mh.systems.clubhouse.ui.adapter.expandableAdapter.CompResultsExpandListAdapter;
 import com.mh.systems.clubhouse.utils.constants.ApplicationGlobal;
 import com.mh.systems.clubhouse.web.api.WebAPI;
 import com.mh.systems.clubhouse.web.models.AJsonParamsResultOfCompetition;
 import com.mh.systems.clubhouse.web.models.CompetitionDetailItems;
 import com.mh.systems.clubhouse.web.models.CompetitionResultAPI;
+import com.mh.systems.clubhouse.web.models.Result;
 import com.mh.systems.clubhouse.web.models.ResultEntries;
 import com.mh.systems.clubhouse.web.api.WebServiceMethods;
 import com.mh.systems.clubhouse.utils.ScrollRecycleView;
@@ -31,6 +37,12 @@ import net.opacapp.multilinecollapsingtoolbar.CollapsingToolbarLayout;
 
 import java.lang.reflect.Type;
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.TreeMap;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
@@ -61,8 +73,6 @@ public class CompletedDetailActivity extends BaseActivity {
     TextView tvEventStatusStrDD;
     @Bind(R.id.nsvContent)
     NestedScrollView nsvContent;
-    @Bind(R.id.tvResultDesc)
-    TextView tvResultDesc;
     @Bind(R.id.tvNoDataView)
     TextView tvNoDataView;
 
@@ -84,8 +94,8 @@ public class CompletedDetailActivity extends BaseActivity {
 
     CompetitionDetailAdapter competitionDetailAdapter;
 
-    @Bind(R.id.lvListOfMembers)
-    ListView lvListOfMembers;
+   /* @Bind(R.id.lvListOfMembers)
+    ListView lvListOfMembers;*/
 
     @Bind(R.id.llRankOfMembers)
     LinearLayout llRankOfMembers;
@@ -98,6 +108,13 @@ public class CompletedDetailActivity extends BaseActivity {
     int iPopItemPos;
 
     Typeface tpRobotoMedium, tfSFUITextSemibold, tfButlerLight;
+
+    @Bind(R.id.elvListOfMembers)
+    ExpandableListView elvListOfMembers;
+    CompResultsExpandListAdapter expListAdapter;
+    List<String> expandableListTitle;
+    Map<String, List<ResultEntries>> listHashMapOfResults = new TreeMap<>();
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -323,11 +340,19 @@ public class CompletedDetailActivity extends BaseActivity {
 
                     tvNoDataView.setVisibility(View.INVISIBLE);
 
-                    tvResultDesc.setText(competitionDetailItems.getCompResultData().getResults().get(0).getDescription());
+                    List<Result> mResultDataList = competitionDetailItems.getCompResultData().getResults();
 
-                    competitionDetailAdapter = new CompetitionDetailAdapter(CompletedDetailActivity.this, resultEntryArrayList);
-                    lvListOfMembers.setAdapter(competitionDetailAdapter);
-                    ScrollRecycleView.getListViewSize(lvListOfMembers);
+                    for (int iCount = mResultDataList.size() - 1; iCount >= 0; iCount--) {
+                        String strKey = mResultDataList.get(iCount).getResultID() + mResultDataList.get(iCount).getDescription();
+                        //Log.e("icount", ""+mResultDataList.get(iCount).getDescription());
+                        listHashMapOfResults.put(strKey, mResultDataList.get(iCount).getResultEntries());
+                    }
+
+                    setupExpandableList();
+
+                    //competitionDetailAdapter = new CompetitionDetailAdapter(CompletedDetailActivity.this, resultEntryArrayList);
+                    //  lvListOfMembers.setAdapter(competitionDetailAdapter);
+                    //  ScrollRecycleView.getListViewSize(lvListOfMembers);
 
                     //Forcefully scroll UP of screen after loading.
                     nsvContent.post(new Runnable() {
@@ -349,6 +374,110 @@ public class CompletedDetailActivity extends BaseActivity {
             reportRollBarException(CompletedDetailActivity.class.getSimpleName(), e.toString());
         }
         hideProgress();
+    }
+
+      private void setupExpandableList() {
+        expandableListTitle = new ArrayList<String>(listHashMapOfResults.keySet());
+        expListAdapter = new CompResultsExpandListAdapter(this, expandableListTitle, listHashMapOfResults);
+        elvListOfMembers.setAdapter(expListAdapter);
+        ScrollRecycleView.getListViewSize(elvListOfMembers);
+        /*elvListOfMembers.setOnGroupExpandListener(new ExpandableListView.OnGroupExpandListener() {
+
+            @Override
+            public void onGroupExpand(int groupPosition) {
+
+                int height = 0;
+                for (int i = 0; i < elvListOfMembers.getChildCount(); i++) {
+                    height += elvListOfMembers.getChildAt(i).getMeasuredHeight();
+                    height += elvListOfMembers.getDividerHeight();
+                }
+                elvListOfMembers.getLayoutParams().height = (height + 6) * 10;
+            }
+        });*/
+
+        elvListOfMembers.setOnGroupClickListener(new ExpandableListView.OnGroupClickListener() {
+
+            @Override
+            public boolean onGroupClick(ExpandableListView parent, View v,
+                                        int groupPosition, long id) {
+                setListViewHeight(parent, groupPosition);
+                return false;
+            }
+        });
+
+        elvListOfMembers.setOnGroupCollapseListener(new ExpandableListView.OnGroupCollapseListener() {
+
+            @Override
+            public void onGroupCollapse(int groupPosition) {
+            }
+        });
+
+        elvListOfMembers.setOnChildClickListener(new ExpandableListView.OnChildClickListener() {
+            @Override
+            public boolean onChildClick(ExpandableListView parent, View v,
+                                        int groupPosition, int childPosition, long id) {
+                return false;
+            }
+        });
+
+        elvListOfMembers.setOnGroupExpandListener(new ExpandableListView.OnGroupExpandListener() {
+            int previousGroup = -1;
+
+            @Override
+            public void onGroupExpand(int groupPosition) {
+                if (groupPosition != previousGroup)
+                    elvListOfMembers.collapseGroup(previousGroup);
+                previousGroup = groupPosition;
+
+             /*   FrameLayout.LayoutParams param = (FrameLayout.LayoutParams) elvListOfMembers.getLayoutParams();
+                param.height = elvListOfMembers.getHeight();
+                elvListOfMembers.setLayoutParams(param);
+                elvListOfMembers.refreshDrawableState();
+
+                int height = 0;
+                for (int i = 0; i < elvListOfMembers.getChildCount(); i++) {
+                    height += elvListOfMembers.getChildAt(i).getMeasuredHeight();
+                    height += elvListOfMembers.getDividerHeight();
+                }
+                elvListOfMembers.getLayoutParams().height = (height + 6) * 5;*/
+            }
+        });
+    }
+
+    private void setListViewHeight(ExpandableListView listView,
+                                   int group) {
+        ExpandableListAdapter listAdapter = (ExpandableListAdapter) listView.getExpandableListAdapter();
+        int totalHeight = 0;
+        int desiredWidth = View.MeasureSpec.makeMeasureSpec(listView.getWidth(),
+                View.MeasureSpec.EXACTLY);
+        for (int i = 0; i < listAdapter.getGroupCount(); i++) {
+            View groupItem = listAdapter.getGroupView(i, false, null, listView);
+            groupItem.measure(desiredWidth, View.MeasureSpec.UNSPECIFIED);
+
+            totalHeight += groupItem.getMeasuredHeight();
+
+            if (((listView.isGroupExpanded(i)) && (i != group))
+                    || ((!listView.isGroupExpanded(i)) && (i == group))) {
+                for (int j = 0; j < listAdapter.getChildrenCount(i); j++) {
+                    View listItem = listAdapter.getChildView(i, j, false, null,
+                            listView);
+                    listItem.measure(desiredWidth, View.MeasureSpec.UNSPECIFIED);
+
+                    totalHeight += listItem.getMeasuredHeight();
+
+                }
+            }
+        }
+
+        ViewGroup.LayoutParams params = listView.getLayoutParams();
+        int height = totalHeight
+                + (listView.getDividerHeight() * (listAdapter.getGroupCount() - 1));
+        if (height < 10)
+            height = 200;
+        params.height = height + 200;
+        listView.setLayoutParams(params);
+        listView.requestLayout();
+
     }
 
 }