+++ MHS CLUBHOUSE/MHSystems/app/src/main/java/com/mh/systems/clubhouse/ui/activites/DashboardActivity.java
@@ -15,7 +15,7 @@ import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.View;
-import android.widget.Button;
+import android.widget.FrameLayout;
 import android.widget.GridView;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
@@ -26,24 +26,25 @@ import com.google.android.gms.common.GoogleApiAvailability;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
+
 import com.mh.systems.clubhouse.R;
+import com.mh.systems.clubhouse.dataaccess.push.QuickstartPreferences;
+import com.mh.systems.clubhouse.dataaccess.push.RegistrationIntentService;
 import com.mh.systems.clubhouse.ui.adapter.RecyclerAdapter.DashboardRecyclerAdapter;
 import com.mh.systems.clubhouse.utils.SyncMarket;
 import com.mh.systems.clubhouse.utils.constants.ApplicationGlobal;
+import com.mh.systems.clubhouse.web.api.WebAPI;
+import com.mh.systems.clubhouse.web.api.WebServiceMethods;
 import com.mh.systems.clubhouse.web.models.deletetoken.AJsonParamsDeleteToken;
 import com.mh.systems.clubhouse.web.models.deletetoken.DeleteTokenAPI;
 import com.mh.systems.clubhouse.web.models.deletetoken.DeleteTokenResult;
-import com.mh.systems.clubhouse.web.models.unreadnewscount.AJsonParamsGetUnreadCount;
-import com.mh.systems.clubhouse.web.models.unreadnewscount.GetUnreadNewsCountAPI;
-import com.mh.systems.clubhouse.web.models.unreadnewscount.GetUnreadNewsResponse;
 import com.mh.systems.clubhouse.web.models.featuresflag.AJsonParamsFeaturesFlag;
 import com.mh.systems.clubhouse.web.models.featuresflag.FeatureFlagsAPI;
 import com.mh.systems.clubhouse.web.models.featuresflag.FeatureFlagsResponse;
+import com.mh.systems.clubhouse.web.models.unreadnewscount.AJsonParamsGetUnreadCount;
+import com.mh.systems.clubhouse.web.models.unreadnewscount.GetUnreadNewsCountAPI;
+import com.mh.systems.clubhouse.web.models.unreadnewscount.GetUnreadNewsResponse;
 import com.mh.systems.clubhouse.web.models.weather.WeatherApiResponse;
-import com.mh.systems.clubhouse.dataaccess.push.QuickstartPreferences;
-import com.mh.systems.clubhouse.dataaccess.push.RegistrationIntentService;
-import com.mh.systems.clubhouse.web.api.WebAPI;
-import com.mh.systems.clubhouse.web.api.WebServiceMethods;
 
 import java.lang.reflect.Type;
 import java.text.SimpleDateFormat;
@@ -70,9 +71,6 @@ public class DashboardActivity extends BaseActivity {
 
     private final int PLAY_SERVICES_RESOLUTION_REQUEST = 9000;
 
-    //Use for version feature on web.
-    private final int TAG_ANDROID = 2;
-
     /*********************************
      * INSTANCES OF CLASSES
      *******************************/
@@ -84,13 +82,10 @@ public class DashboardActivity extends BaseActivity {
     LinearLayout llLogoutBtn;
 
     @Bind(R.id.llSettings)
-    LinearLayout llSettings;
+    FrameLayout llSettings;
 
     @Bind(R.id.llWeatherGroup)
     LinearLayout llWeatherGroup;
-
-    @Bind(R.id.btSendFeedback)
-    Button btSendFeedback;
 
     @Bind(R.id.tvTodayTemperature)
     TextView tvTodayTemperature;
@@ -144,87 +139,17 @@ public class DashboardActivity extends BaseActivity {
      *******************************/
     ArrayList<DashboardItems> dashboardItemsArrayList = new ArrayList<>();
 
-    int iHandicapPosition = 0;
+    int iHandicapPosition = -1;
     String strNameOfWeatherLoc = "";
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_dashboard);
+        //setContentView(R.layout.activity_dashboard_nine_items);
 
-        /**
-         * Annotate fields with @Bind and a view ID for Butter Knife to find and
-         * automatically cast the corresponding view in your layout.
-         */
-        ButterKnife.bind(DashboardActivity.this);
-
-        try {
-            checkUpdateVersion();
-        } catch (PackageManager.NameNotFoundException e) {
-            e.printStackTrace();
-        }
-
-        //Initialize adapter.
-        dashboardRecyclerAdapter = new DashboardRecyclerAdapter(this,
-                dashboardItemsArrayList,
-                iHandicapPosition,
-                loadPreferenceValue(ApplicationGlobal.KEY_HCAP_EXACT_STR, "N/A"));
-        gvMenuOptions.setAdapter(dashboardRecyclerAdapter);
-
-        sendTokenToServer();
-
-        //LogOut listener.
-        llLogoutBtn.setOnClickListener(mLogoutListener);
-
-        //LogOut listener.
-        llLogoutBtn.setOnClickListener(mLogoutListener);
-
-        //Settings click event handle here.
-        llSettings.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                intent = new Intent(DashboardActivity.this, SettingsActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        //Send Feedback click event here.
-        btSendFeedback.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                intent = new Intent(DashboardActivity.this, SendFeedbackActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        //See the weather of 5 days
-        llWeatherGroup.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                intent = new Intent(DashboardActivity.this, WeatherDetailActivity.class);
-                intent.putExtra("WEATHER_LOC", strNameOfWeatherLoc);
-                startActivity(intent);
-            }
-        });
-//        String temp = loadPreferenceValue(ApplicationGlobal.KEY_TEMPKEY_TEMPERATURE, "");
-//          if (temp.equals("")){
-//        callWeatherService();
-//          }
-//    else{
-//            llWeatherGroup.setVisibility(View.VISIBLE);
-//            tvTodayTemperature.setText(loadPreferenceValue(ApplicationGlobal.KEY_TEMPKEY_TEMPERATURE, ""));
-//            tvWeatherDesc.setText(loadPreferenceValue(ApplicationGlobal.KEY_TEMPKEY_WEATHER, ""));
-//            //        tvNameOfLocation.setText(weatherData.getName() + ", " + weatherData.getSys().getCountry());
-//            tvNameOfLocation.setText(loadPreferenceValue(ApplicationGlobal.KEY_TEMPKEY_LOCATION, ""));
-//        //    todayIcon.setImageURI(Uri.parse("http://openweathermap.org/img/w/" + weatherData.getWeather().get(0).getIcon() + ".png"));
-//            Resources res=getResources();
-//            int resID = res.getIdentifier(loadPreferenceValue(ApplicationGlobal.KEY_TEMPKEY_IMAGE, ""), "mipmap", getPackageName());
-//            Drawable drawable = res.getDrawable(resID);
-//            todayIcon.setImageDrawable(drawable);
-//
-//        }
+        //Dashboard is dynamic and enable/disable by Feature flag.
     }
-
 
     @Override
     protected void onResume() {
@@ -232,7 +157,7 @@ public class DashboardActivity extends BaseActivity {
 
         if (isOnline(DashboardActivity.this)) {
             callFeaturesFlagService();
-            callWeatherService();
+            //callWeatherService();
         }
     }
 
@@ -299,6 +224,24 @@ public class DashboardActivity extends BaseActivity {
         }
     }
 
+
+    /**
+     * Logout user from app and navigate back to
+     * Login screen.
+     */
+    public View.OnClickListener mLogoutListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+
+            //Call Weather api functionality.
+            if (isOnline(DashboardActivity.this)) {
+                callLogoutService();
+            } else {
+                showAlertMessage(getString(R.string.error_no_connection));
+            }
+        }
+    };
+
     /**
      * Check if any update version available on
      * Google Store.
@@ -347,6 +290,7 @@ public class DashboardActivity extends BaseActivity {
      * Implements this method to send Token to
      * server for push notifications.
      */
+
     private void sendTokenToServer() {
         mRegistrationBroadcastReceiver = new BroadcastReceiver() {
             @Override
@@ -364,24 +308,6 @@ public class DashboardActivity extends BaseActivity {
             startService(intent);
         }
     }
-
-    /**
-     * Logout user from app and navigate back to
-     * Login screen.
-     */
-    private View.OnClickListener mLogoutListener = new View.OnClickListener() {
-        @Override
-        public void onClick(View v) {
-
-            //Call Weather api functionality.
-            if (isOnline(DashboardActivity.this)) {
-                callLogoutService();
-            } else {
-                showAlertMessage(getString(R.string.error_no_connection));
-            }
-        }
-    };
-
 
     /**
      * Implements a method to set Grid MENU options
@@ -414,13 +340,6 @@ public class DashboardActivity extends BaseActivity {
                     getApplicationContext().getPackageName() + ".ui.activites.CourseDiaryActivity"));
         }
 
-        /*//Pro-Agenda Book Lessons
-        dashboardItemsArrayList.add(new DashboardItems(
-                R.mipmap.ic_booking_agenda,
-                "Book Lessons",
-                getApplicationContext().getPackageName() + ".ui.activites.CourseDiaryWebviewActivity"));*/
-
-
         //Add Competitions
         if (loadPreferenceBooleanValue(ApplicationGlobal.KEY_COMPETITIONS_FEATURE, false)) {
 
@@ -428,6 +347,14 @@ public class DashboardActivity extends BaseActivity {
                     R.mipmap.ic_home_competitions,
                     "Competitions",
                     getApplicationContext().getPackageName() + ".ui.activites.CompetitionsActivity"));
+        }
+
+        //Tee Time Booking Integration
+        if (loadPreferenceBooleanValue(ApplicationGlobal.KEY_MOTT_FEATURE, false)) {
+            dashboardItemsArrayList.add(new DashboardItems(
+                    R.mipmap.ic_teetime_booking,
+                    "Tee Time Booking",
+                    getApplicationContext().getPackageName() + ".ui.activites.TeeTimeBookingActivity"));
         }
 
         //Add Members
@@ -454,12 +381,71 @@ public class DashboardActivity extends BaseActivity {
                     getApplicationContext().getPackageName() + ".ui.activites.YourAccountActivity"));
         }
 
-        //Set Grid options adapter.
-        dashboardRecyclerAdapter.notifyDataSetChanged();
+        /*dashboardRecyclerAdapter = new DashboardRecyclerAdapter(this,
+                dashboardItemsArrayList,
+                iHandicapPosition,
+                loadPreferenceValue(ApplicationGlobal.KEY_HCAP_EXACT_STR, "N/A"));
+        gvMenuOptions.setAdapter(dashboardRecyclerAdapter);
+        dashboardRecyclerAdapter.notifyDataSetChanged();*/
+
+        if (dashboardItemsArrayList.size() <= 6) {
+            setContentView(R.layout.activity_dashboard_six_items);
+        } else {
+            setContentView(R.layout.activity_dashboard_nine_items);
+        }
+
+        initViewResources();
+        callWeatherService();
 
         setupGridLayout(dashboardItemsArrayList.size());
 
         // ScrollRecycleView.getListViewSize(gvMenuOptions);
+    }
+
+    private void initViewResources() {
+        /**
+         * Annotate fields with @Bind and a view ID for Butter Knife to find and
+         * automatically cast the corresponding view in your layout.
+         */
+        ButterKnife.bind(DashboardActivity.this);
+
+        try {
+            checkUpdateVersion();
+        } catch (PackageManager.NameNotFoundException e) {
+            e.printStackTrace();
+            reportRollBarException(DashboardActivity.class.getSimpleName(), e.toString());
+        }
+
+        //Initialize adapter.
+        dashboardRecyclerAdapter = new DashboardRecyclerAdapter(this,
+                dashboardItemsArrayList,
+                iHandicapPosition,
+                loadPreferenceValue(ApplicationGlobal.KEY_HCAP_EXACT_STR, "N/A"));
+        gvMenuOptions.setAdapter(dashboardRecyclerAdapter);
+
+        sendTokenToServer();
+
+        //LogOut listener.
+        llLogoutBtn.setOnClickListener(mLogoutListener);
+
+        //Settings click event handle here.
+        llSettings.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                intent = new Intent(DashboardActivity.this, SettingsActivity.class);
+                startActivity(intent);
+            }
+        });
+
+        //See the weather of 5 days
+        llWeatherGroup.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                intent = new Intent(DashboardActivity.this, WeatherDetailActivity.class);
+                intent.putExtra("WEATHER_LOC", strNameOfWeatherLoc);
+                startActivity(intent);
+            }
+        });
     }
 
     /**
@@ -498,7 +484,7 @@ public class DashboardActivity extends BaseActivity {
                 });
                 break;
 
-            default:
+            case 6:
                 layoutManager.setSpanSizeLookup(new GridLayoutManager.SpanSizeLookup() {
                     @Override
                     public int getSpanSize(int position) {
@@ -506,14 +492,31 @@ public class DashboardActivity extends BaseActivity {
                     }
                 });
                 break;
+
+            case 9:
+            case 7:
+                layoutManager.setSpanSizeLookup(new GridLayoutManager.SpanSizeLookup() {
+                    @Override
+                    public int getSpanSize(int position) {
+                        return position == 3 ? 6 : 2;
+                    }
+                });
+                break;
+
+            default:
+                layoutManager.setSpanSizeLookup(new GridLayoutManager.SpanSizeLookup() {
+                    @Override
+                    public int getSpanSize(int position) {
+                        return position == 3 ? 6 : 2;
+                    }
+                });
+                break;
         }
-
-
-        // int spacingInPixels = getResources().getDimensionPixelSize(R.dimen.spacing);
-        // gvMenuOptions.addItemDecoration(new SpacesItemDecoration(spacingInPixels));
 
         // Layout Managers:
         gvMenuOptions.setLayoutManager(layoutManager);
+//         int spacingInPixels = 2;//getResources().getDimensionPixelSize(R.dimen.spacing);
+//         gvMenuOptions.addItemDecoration(new SpacesItemDecoration(spacingInPixels));
     }
 
     /**
@@ -521,7 +524,7 @@ public class DashboardActivity extends BaseActivity {
      * Because above two grid items should be in center of below
      * three one.
      */
-   /* public class SpacesItemDecoration extends RecyclerView.ItemDecoration {
+   /* private class SpacesItemDecoration extends RecyclerView.ItemDecoration {
         private int space;
 
         public SpacesItemDecoration(int space) {
@@ -605,12 +608,8 @@ public class DashboardActivity extends BaseActivity {
             Drawable drawable = ContextCompat.getDrawable(DashboardActivity.this, resID);
             todayIcon.setImageDrawable(drawable);
 
-//            savePreferenceValue(ApplicationGlobal.KEY_TEMPKEY_TEMPERATURE, ("" + ((int) (weatherData.getMain().getTemp() - 273.15f)) + "°C"));
-//            savePreferenceValue(ApplicationGlobal.KEY_TEMPKEY_WEATHER, ("Today, "+(desc.substring(0, 1).toUpperCase() + desc.substring(1))));
-//            savePreferenceValue(ApplicationGlobal.KEY_TEMPKEY_LOCATION, weatherData.getName());
-//            savePreferenceValue(ApplicationGlobal.KEY_TEMPKEY_IMAGE, ("e"+weatherData.getWeather().get(0).getIcon()));
-
         } else {
+			llWeatherGroup.setVisibility(View.GONE);
             Log.e(LOG_TAG, weatherApiResponse.getMessage());
         }
     }
@@ -795,7 +794,6 @@ public class DashboardActivity extends BaseActivity {
         aJsonParamsFeaturesFlag = new AJsonParamsFeaturesFlag();
         aJsonParamsFeaturesFlag.setCallid(ApplicationGlobal.TAG_GCLUB_CALL_ID);
         aJsonParamsFeaturesFlag.setVersion(ApplicationGlobal.TAG_GCLUB_VERSION);
-        aJsonParamsFeaturesFlag.setClubHouseApp(true);
 
         featureFlagsAPI = new FeatureFlagsAPI(getClientId(), "GETCLUBFEATURES", aJsonParamsFeaturesFlag, "CLUBINFO", ApplicationGlobal.TAG_GCLUB_MEMBERS);
 
@@ -817,8 +815,6 @@ public class DashboardActivity extends BaseActivity {
             @Override
             public void failure(RetrofitError error) {
                 Log.e(LOG_TAG, "RetrofitError : " + error);
-//                showAlertMessage(error.getMessage());
-                onBackPressed();
                 hideProgress();
             }
         });
@@ -844,14 +840,19 @@ public class DashboardActivity extends BaseActivity {
              */
             if (featureFlagsResponse.getMessage().equalsIgnoreCase("Success")) {
 
-                //Make Dashboard dynamic according these bool values.
+                 //Make Dashboard dynamic according these bool values.
                 savePreferenceBooleanValue(ApplicationGlobal.KEY_COURSE_DIARY_FEATURE, featureFlagsResponse.getData().getCourseDiaryFeatures());
                 savePreferenceBooleanValue(ApplicationGlobal.KEY_COMPETITIONS_FEATURE, featureFlagsResponse.getData().getCompetitionsFeature());
                 savePreferenceBooleanValue(ApplicationGlobal.KEY_HANDICAP_FEATURE, featureFlagsResponse.getData().getHandicapFeature());
                 savePreferenceBooleanValue(ApplicationGlobal.KEY_MEMBERS_FEATURE, featureFlagsResponse.getData().getMembersFeature());
                 savePreferenceBooleanValue(ApplicationGlobal.KEY_CLUB_NEWS_FEATURE, featureFlagsResponse.getData().getClubNewsFeature());
                 savePreferenceBooleanValue(ApplicationGlobal.KEY_YOUR_ACCOUNT_FEATURE, featureFlagsResponse.getData().getYourAccountFeature());
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_MOTT_FEATURE, featureFlagsResponse.getData().getMOTTFeature());
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_MY_EVENT_FEATURE, featureFlagsResponse.getData().isMyEventFeature());
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_MY_EVENT_ONLY, featureFlagsResponse.getData().isMyEventOnly());
 
+                savePreferenceValue(ApplicationGlobal.KEY_GENDER_FILTER, featureFlagsResponse.getData().getGenderFilter());
+				
                 setGridMenuOptions();
 
             } else {