+++ MHS CLUBHOUSE/MHSystems/app/src/main/java/com/mh/systems/clubhouse/ui/adapter/BaseAdapter/CompetitionsAdapter.java
@@ -163,6 +163,7 @@ public class CompetitionsAdapter extends BaseAdapter {
         viewHolder.llCompetitionGroup.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+
                 Intent intent = null;
 
                 if (iPopItemPos <= 1) {