+++ MHS CLUBHOUSE/MHSystems/app/src/main/java/com/mh/systems/clubhouse/utils/constants/ApplicationGlobal.java
@@ -36,6 +36,7 @@ public class ApplicationGlobal {
     public static final int POSITION_MEMBERS = 4;
     public static final int POSITION_ABOUT_CLUB = 5;
     public static final int POSITION_MY_ACCOUNT = 6;
+    public static final int POSITION_TEE_TIME_BOOKING = 8;
 
     /**
      * Use this for COMPETITIONS ENTRY.
@@ -123,6 +124,10 @@ public class ApplicationGlobal {
     public static final String KEY_CLUB_NEWS_FEATURE = "ClubNewsFeature";
     public static final String KEY_YOUR_ACCOUNT_FEATURE = "YourAccountFeature";
 
+    public static final String KEY_MOTT_FEATURE = "MOTTFeature";
+    public static final String KEY_MY_EVENT_FEATURE = "MyEventFeature";
+    public static final String KEY_MY_EVENT_ONLY = "MyEventOnly";
+    public static final String KEY_GENDER_FILTER = "GenderFilter";
 
     /**
      * SharedPrefernce Key for weather store