+++ MHS CLUBHOUSE/MHSystems/app/src/main/java/com/mh/systems/clubhouse/web/models/featuresflag/AJsonParamsFeaturesFlag.java
@@ -12,10 +12,12 @@ public class AJsonParamsFeaturesFlag {
     @SerializedName("callid")
     @Expose
     private String callid;
-
-    @SerializedName("ClubHouseApp")
+    @SerializedName("memberid")
     @Expose
-    private boolean ClubHouseApp;
+    private String memberid;
+    @SerializedName("LoginMemberId")
+    @Expose
+    private String LoginMemberId;
 
     public Integer getVersion() {
         return version;
@@ -33,12 +35,19 @@ public class AJsonParamsFeaturesFlag {
         this.callid = callid;
     }
 
-
-    public boolean getClubHouseApp() {
-        return ClubHouseApp;
+    public String getMemberid() {
+        return memberid;
     }
 
-    public void setClubHouseApp(boolean clubHouseApp) {
-        ClubHouseApp = clubHouseApp;
+    public void setMemberid(String memberid) {
+        this.memberid = memberid;
+    }
+
+    public String getLoginMemberId() {
+        return LoginMemberId;
+    }
+
+    public void setLoginMemberId(String LoginMemberId) {
+        this.LoginMemberId = LoginMemberId;
     }
 }