+++ MHS CLUBHOUSE/MHSystems/CHANGELOG.md
@@ -1,3 +1,12 @@
+####Version 1.1 ()
+
+  ####24th Nov, 2017
+  - Competitions : Add Option to choose which competitions to display
+  - Competition Results : Display All results in collapsible way 
+  - Mobile App Item 131: URGENT BUG. Handicap Info not Correct
+
 ####Version 1.0 (18th Aug, 2017)
 
 Initial release
+
+